+++ CHANGELOG.md
@@ -1,8 +1,17 @@
+0.5 (unreleased)
+================
+- Support parallel reads in `open_ncml` using `dask`. By @huard
+
+
 0.4 (unreleased)
 ================
 
 - Add support for <EnumTypeDef>. By @bzah
-- Support parallel reads in `open_ncml` using `dask`. By @huard
+- Update XSD schema and dataclasses to latest version from netcdf-java to add support
+  for unsigned types. By @bzah
+- Add support for scalar variables. By @Bzah
+- [fix] empty attributes now are parsed into an empty string instead of crashing the parser.  By @Bzah
+
 
 0.3.1 (2023-11-10)
 ==================